+++ CHANGELOG.md
@@ -10,6 +10,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [2.1.0] - 2020-12-26
+
+### Added
+- feat: Add alerts providers (#114)
+
+### Changed
+- chore(deps): Update dependencies
+
+### Fixed
+- docs: Fix some readme typos
+
+### Removed
+- fix: Remove mocks-server-plugin tag from package.json
+
 ## [2.0.5] - 2020-12-21
 
 ### Added